Low Level Task Widget UI
Ref #50

Updated the UI of the Task Widget to scale
properly and be a little more elegant.
Also moved from using a 9 slice texture to a panel
with rounded edges for the background. This solution
is a lot more workable and flexible.
The default view port settings have also been set to 300 x 800
which is the most common viewport resolution for mobile.
This should help with making the UI.
Looking into some sort of way to make the border-radius
of elements consistent with screen size.