Added Medium Level Task Widget
Ref #50

Added a Label to show the Task Info at medium levl of detail.

Hopefully all the task widgets will be able to be
represented in one scene. The idea being that when
scaled everything should fall into place, and this will
be controlled by a script. To reflect this idea the name
of some of the UI elements have been tagged with Low and Med
to signify what level of detail they will be present for.