Level of Detail Script
Ref #50

Added a script to the Task Widget that allow
the widget to switch between levels of detail.
It's very rudeamentary and not the nicest code at
the moment.
The widget also isn't as flexible in its scaling as
I would like but it is perfectly acceptable for a
prototype level at the moment.
Will continue to add exported properties that allow
information on the widget to be changed, but can only
do so much until we decide on how we are storing the
task data.