Text and Image properties for Task Widget
Ref #50

The text and image of tasks can now be set in the inspector
this will update the information when the game is ran.
An effect of this is I found out that there is a little
scaling issue with certain length of text, and there
probably ought to be some sort of minimum size property on
task info text in the High Level Detail version, aswell as
some truncation if the text is too long.