Skeleton Layout for Details Page
Refs #60

Created a basic layout for the UI of the details page. It is semi
dynamic, as in there is some scaling but there is no easy way to input
the data and have it up date.

So the next steps will be to allow the page to be updated with new data
easily.