Staff on Task Node
Refs #60

Added custom Godot nodes for Staff Icon and Staff On Task, this will
allow these UI elements to be dynamically changed and used elsewhere.

Currently these values are just changed in the inspector but when fully
implemented with the rest of the program it should be sent the data from
the task that was clicked.

Currently don't have a way to change photos from the inspector, so just
uses the default photo only for now.