Themes and Scope Change
Refs #60

Added themes to the UI so its easier to repeat.
Some extra clean up with some unused script that are no longer needed
for the UI.
Changing scope, so that staff on task and task log are currently not
needed in the task page. This should greatly increase the chance of
getting a working prototype by the next sprint.