Dynamic Resource UI
Refs #60

Resources now have a specific UI component that allows them to be
dynamically updated with a resource dictionary. It will auto size and
auto layout all the resources sent in. Not fully complete because we
dont have a resource manager.
Also connected this up with the task widgets, so clicking on a task
widget for details will show the correct resources on the detail page.