Task Widgets Displaying Properly on Map
Refs #50

Task widgets spawn in the correct position on the map according to their
task data.
They also resize correctly when zoomed in.
They also change level of detail to medium at a certain threshold of
zoom.
There is currently an issue with the scaling of the high level of detail
that needs to be fixed.
In the future this will also need to be hooked up to some sort of task
manager as a place to get the task data from.