Relationships update on end of task
Refs #72

Relationships with senders will now update when a task is finished.
There is support for updating relationship when a task is ended early,
and ending a task early can give negative relationships.

There is probably a need to do some general clean up within the code and
a lot of things are tied together that shouldn't be.
Ending a task early is also a priority