Relationship UI and Fixes
Refs #72

Added UI to show the sender and their relationship in the message UI.
The relationship bar changes color depending on the relationship and
there is also a tooltip which gives one word for the nature of the
relationship e.g "strained", "excellent" or "neutral...

Added two temporary senders for testing, government and UN OCHA agent.

Fixed some issues where the relationship was not actually getting
applied to the sender.

Need to add some back up to the sender UI incase no sender is sent in to
default to some default properties.