Added ReceivedMessage UI Element
Refs #91

There is now a UI element that shows a message that is sent, the turns
left to answer on that message and the contact information. Switching
between states of the message i.e Unread and Read updates the
information accordingly.
It might be worth looking into some sort of MessageInstance resource to
store this data rather than the UI element to keep the data and UI
seperate.
The UI element could also be prettified later on, maybe take in a font
for the answer now colour so it can be bold too.
Also did not add the arrow that appears on the design as it felt too
cluttered and was difficult to get the look correct.