Added Message Instance Resource
Refs #91

Resource for message instance that can keep track of the message status
and the current number of turns remaining for a message. Will need to
reflect this in the documentation but also the message manager should
send out message instances rather than an actual message. Will need to
change this later.