Message Page
Refs #94

There is now a message page that shows a message and a player response
and the contact details. This is not yet hooked up to a response page or
the messages received paged but will do that next.
It currently works under the assumption that a conversation will go no
further than question -> response. The message renderer could handle a
longer conversation but the data is not set up to store that.