Correct Message Selection
Refs #91 #94

Selecting a message in the message received page now shows the correct
message. It will also update the message status to be READ, which will
get ride of the unread badge on the message received page.
Might be cleaner to set some of this up with an Event Bus of sorts but
will see how it goes