Message Response Page
Refs #93

Added a page which will allow the player to make a response. Also made
some themes that can be reused elsewhere in the UI so everything can be
updated in one place and will lead to a more consistent color scheme.
Still need to hook this up to the message received page etc and allow it
to actually start tasks.