Finished UI of New Message System
Closes #94 #93 #91

Fully implemented the UI pages for messages. This included a new
EventBus system to prevent coupling and backward references. Adding a
theme to buttons for consistency. Changing the response resource to
include a title.
For future notes: a task manager would be ideal to help decouple the UI
and logic in map_tasks
There is still a bug with being able to select a task that will get rid
of your resources.
It could still be beneficial to add message filters.