Tasks Cancelling Behaviour
Refs #88

Added 5 different behaviour types for cancelling that are documented in
the code and will also be put in the documentation. Fixed a small bug
with the prerequisites (it was treating prereqs as antireqs). Really do
need to work on this task manager because map_task is really quite a
mess. Also added a default response option.