Task Manager
Refs #98

Added a task manager singleton that keeps track of current tasks and
updates them.
Currently exploring the idea of reducing the number of singletons and
therefore coupling. Message Manager would be very easy to remove as a
singleton and could be done now with the introduction of the EventBus.
The one piece of coupling I can't really figure out atm is to do with
the prerequisite validations. Maybe some sort of game manager that holds
task and message manager and will pass the updated tasks into message
manager.