Task Resource Validation
Refs #96

If there is not enough resources the player may not confirm the option.
It also highlights what resources the player is low on. Edited the
task_resources screen so is similar to the increments but can choose to
show without arrows. Was easier to add than I thought it'd be.