Fixes to Message Renderering
Some small fixes before writing the tests for the
message renderer. Was not rendering the player
messages in the best way and I thought of a cleaner
way to do it. Also added a no messages label on the
message received page that shows when the player has
no new messages.