Fake Merge with Message System
Refs #102

Would be too much effort to rebase or merge to message system at the
moment and I dont exactly need all the features in message system so
just copied what I needed. A little janky, probably a nicer way to do it
but for right now alot easier for everyone involved.