Started work on an example Scenario
Refs #102

Tried using the tooling to make a scenario, is quicker than doing it
with making resources but there is still a few issues that need solved
to make it way better. One such issue is the arranging of nodes, I think
I've found a good solution to that so will start work on it now, then
rewrite the codebase a little.