Graph Sorted on Load
Refs #102

It's not perfect as usually the prerequsites will get crossed over but
thats very easy to sort out. There is no auto sort button anymore that
didn't work, wanted t osee if there was some sort of other arranging
algorithm that could be implemented but really not worth the time. y