Fix Lambda Capture Error
Found a small bug that was throwing a lambda capture error when a task
had been compelted, and then another task was completed. This was
because a signal was not being disconnected once it was fired, rewrote
how this works so the error isn't there anymore. Error didn't crash the
program but it is good to not have it there anymore.