FIX: Relationship not being applied on 0 turn task
Fixes #103

There was an issue with relationships not being applied when a task took
0 turns, which was particularly important for these kinds of tasks as
they were usually the tasks that would drop a relationship status.

Removed relationship manager from applying the relationships as this
leads to some undefined behaviour in regards to script execution order.

Fixed this by changing how the task manager deals with tasks with 0
time.