Updating Scenario
Refs #81

Added new tasks to the scenario. Should be up to date to the
Canva scenario and Tims feedback, minus the mudslide which has
been removed because I don't like it. No idea how the balancing is
and need to pick positions for the tasks. Will try it out in a bit.