Updated Scripts to Use Time String Function
Refs #113

Any place that uses a time string now uses the function that is in
Global Timer. Also updated the program to run on turns only, so task
completion is measured in turns rather than minutes or hours. Means that
the timescale of the scenario can be changed more flexibly. Important to
note though that alot of unit tests run on the expecatation that the
time step will be 60.