Fixed Multiple Line Rendering Issues
Closes #110

Fixed an issue where if a message had a line break, it would render
weirdly in the received message page. The new line is now stripped
before rendering and the maximum lines of text in the message preview
is set to 1.

Tests have been added to ensure there is no regression.