Animations for Task Widgets
Refs #145

Added some animations for clicking on a widget and for when a widget
finishes.
These are done with tweens. The finishing animation might be overkill
and needs a few more tweaks so is just a prototype for now.
I think the clicking animation is pretty successful however, and feels
fun to click.
The animations are very easy to turn off and are separated from any
logic.
Adding the animation has uncovered that the task widget UI isn't the
easiest to work with and could be improved.