Fixing Task Manager Issues
Refs #146

Fixed all the bugs in the aforementioned issue. A lot of these bugs
ended up being connected in some way. While I was at it, I also cleaned
up the TaskManager code a little and added some documentation.
Changed how tasks were being cancelled, and made it so that only
TaskManager emits the task_finished signal, and then anywhere that wants
to cancel a task can use task_cancelled instead. This leads to much more
easily defined behaviour, as a task being finished is more concrete, and
a task being cancelled is more concrete too.
Most of the issues came from the task instances not been marked as
completed when they entered the completed tasks list, which means they
were then failing the prerequisite check. This was an easy fix, but we
could maybe consider an assumption that if a task enters the completed
tasks list it is assumed to be completed.
Still not a fan of how tasks are being cancelled but I can't think of a
better way at the moment.