Balancing, Tasks and Changes
Refs #81

Changed balancing so the game is much harder and extended some
tasks to have more obvious and likely consequences. Added some
tasks which always happen on the same turn so we can force
decisions at those points, for example do we save the school or
the president, or both, but at what cost.