Updated Task Details Screen
Refs #156

Updated the task details screen to be a bit more 'fun', and to be a bit
more readable. It now shows the sender, shows the resources a bit
better, and has some animations to give a bit more feedback. The page
also updates on a timer progress.
Also updated the resource UI entry to look a bit different, and to have
more exposed visual parameters, like text color, icon color etc. There
is also an option to place a circle behind it.
Fixed some tests with the message page that were broken, and changed it
so the back buttons go back a page rather than closing the whole thing.