Navbar Button Animations
Refs #154

There's now little animations on the navbar button, this includes
a pop in animation for the bubble, and a shake on the icon when
there's unread messages and the page isnt open.
Also changed the navbar buttons to have empty styleboxes to get rid
of the outline that appears around them.
Got rid of the new message button at the top as I feel like the navbar
catches enough attention now.
Also fixed a bug where if all the messages were read in a turn then
a new message was sent the bubble would stay not visible.
Will write tests for all these once team confirms changes.