Design Changes
Refs #158

Changed the program to work in a way that no longer relies on
singletons. This was done by combining the Event Bus and the Scenario
Manager into one script called GameManager, which holds the current
game. Objects now also don't directly reference TaskManager and are
instead passed a task manager, which aids in modularity and reducing
coupling. There is a worry for how many changes this makes but we'll
see.